+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-02 12:55:15
-// Commit: docs(ui): update readme and API contracts for ui
+// Updated at: 2026-04-04 13:29:32
+// Commit: test(ui): implement widget integration tests for ui
 
 class uiComponentState {
-  final String commitRef = '54';
-  final String lastMessage = 'docs(ui): update readme and API contracts for ui';
-  final DateTime timestamp = DateTime.parse('2026-04-02T12:55:15+05:00');
+  final String commitRef = '107';
+  final String lastMessage = 'test(ui): implement widget integration tests for ui';
+  final DateTime timestamp = DateTime.parse('2026-04-04T13:29:32+05:00');
 }